scrapped/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.14.1
google-api-python-client==2.118.0
stripe==7.10.0
requests==2.31.0
python-jose==3.3.0
PyJWT==2.8.0
